+++ src/4.3_to_5.0/__test_resources__/emptyLegacyDrillthrough.ts
