+++ src/4.3_to_5.0/__test_resources__/legacyDrillthrough.ts
@@ -1,4 +1,4 @@
-import type { LegacyWidgetState } from "../4.3_to_5.0/migration.types";
+import type { LegacyWidgetState } from "../migration.types";
 
 /**
  * The widgetState of a legacy drillthrough, useful for unit tests.
