+++ src/4.3_to_5.0/migrateUIFolder43To50.test.ts
@@ -1,14 +1,14 @@
 import _map from "lodash/map";
 import _some from "lodash/some";
 import { migrateUIFolder43To50 } from "./migrateUIFolder43To50";
-import { smallLegacyUIFolder } from "../__test_resources__/smallLegacyUIFolder";
-import { legacyUIFolder } from "../__test_resources__/legacyUIFolder";
-import { servers } from "../__test_resources__/servers";
+import { smallLegacyUIFolder } from "./__test_resources__/smallLegacyUIFolder";
+import { legacyUIFolder } from "./__test_resources__/legacyUIFolder";
+import { servers } from "./__test_resources__/servers";
 import { ContentRecord } from "@activeviam/activeui-sdk";
 import { LegacyDashboardState } from "./migration.types";
-import { smallLegacyPivotFolder } from "../__test_resources__/smallLegacyPivotFolder";
-import { smallLegacyUIFolderWithInvalidFilter } from "../__test_resources__/smallLegacyUIFolderWithInvalidFilter";
-import { smallLegacyUIFolderWithInvalidDashboard } from "../__test_resources__/smallLegacyUIFolderWithInvalidDashboard";
+import { smallLegacyPivotFolder } from "./__test_resources__/smallLegacyPivotFolder";
+import { smallLegacyUIFolderWithInvalidFilter } from "./__test_resources__/smallLegacyUIFolderWithInvalidFilter";
+import { smallLegacyUIFolderWithInvalidDashboard } from "./__test_resources__/smallLegacyUIFolderWithInvalidDashboard";
 
 /**
  *  Returns whether `contentRecord` has a descendant with the id `recordId`.
